admin tjek på CRUD mutations

+++ graphql/resolvers/TeamMutation.ts
@@ -1,9 +1,12 @@
-import { Args, Team } from "../../types/types";
+import { Args, MyContext, Team } from "../../types/types";
 import TeamModel from '../../mongoose/models/teamModel'
+import AppErr from "../../utility/AppError";
 
 
 export default {
-    createTeam: async (_parent:never, { input }: Args) => {
+    createTeam: async (_parent:never, { input }: Args,  {user}: MyContext) => {
+      if(user?.role === "Admin")
+        {
         if('captain' in input){
           let newTeam: Team = {
             name: input.name ? input.name : "",
@@ -15,15 +18,29 @@ export default {
         } else {
           return null;
         }
+      }
+
+      throw new AppErr("Not Authorized", 401)
       },
-      deleteTeam: async (_parent:never, { id }:Args) => {
-        let deletedTeam = await TeamModel.findByIdAndDelete(id)
-        if (deletedTeam === null) {
-          return false; 
+      deleteTeam: async (_parent:never, { id }:Args, {user}: MyContext) => {
+
+        if(user?.role === "Admin")
+        {
+          
+          
+          let deletedTeam = await TeamModel.findByIdAndDelete(id)
+          if (deletedTeam === null) {
+            return false; 
+          }
+          return true;
         }
-        return true;
+
+        throw new AppErr("Not Authorized", 401)
     },
-    updateTeam: async (_parent: never, { id, input }:Args) => {
+    updateTeam: async (_parent: never, { id, input }:Args, {user}: MyContext) => {
+      if(user?.role === "Admin")
+        {
+
         if('captain' in input){
         let team = await TeamModel.findByIdAndUpdate(id, input, {
             new:true,
@@ -31,6 +48,8 @@ export default {
         })
           return team; 
         }
+      }
+      throw new AppErr("Not Authorized", 401)
     }
 
 }
